+++ package.json
@@ -3,7 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.5.0",
     "@hookform/resolvers": "^2.8.3",
+    "@mui/icons-material": "^5.0.5",
+    "@mui/material": "^5.0.6",
     "@reduxjs/toolkit": "^1.6.2",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
@@ -15,7 +18,7 @@
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.18.1",
     "react-redux": "^7.2.6",
-    "react-router-dom": "^6.0.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-tabs": "^3.2.3",
     "recharts": "^2.1.6",
@@ -60,7 +63,7 @@
     "redux-devtools-extension": "^2.13.9"
   },
   "lint-staged": {
-    "*.js": "eslint --cache --fix",
-    "*.{js,css,md}": "prettier --write"
+    "*.{js,jsx}": "eslint --cache --fix",
+    "*.{js,jsx,css,md}": "prettier --write"
   }
 }
