+++ src/components/Container/Container.jsx
@@ -1,3 +1,4 @@
+import s from './Container.module.css';
 
 function Container({ children }) {
   return <div className={s.container}>{children}</div>;
