+++ src/components/PrivateRout.jsx
@@ -1,10 +1,10 @@
-import { Route, Redirect } from 'react-router-dom';
+import { Route } from 'react-router-dom';
 
 function PrivateRoute({ children, ...routeProps }) {
   // const isLoggedIn  = если пользователь  залогинен
   return (
     <Route {...routeProps}>
-      {isLoggedIn ? children : <Redirect to="/login" />}
+      {/* {isLoggedIn ? children : <Redirect to="/login" />} */}
     </Route>
   );
 }
